--- urbackupserver/www2/src/features/status/useRemoveClientsMutation.ts
@@ -1,13 +1,0 @@
-import { useMutation, useQueryClient } from "react-query";
-
-import { urbackupServer } from "../../App";
-
-export function useRemoveClientsMutation() {
-  const queryClient = useQueryClient();
-
-  return useMutation(urbackupServer.removeClients, {
-    onSuccess: () => {
-      queryClient.invalidateQueries("status");
-    },
-  });
-}
